dsp-platform/backend/requirements.txt
# DSP Platform Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Cache & Task Queue
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# OAuth2
authlib==1.2.1
requests==2.31.0
requests-oauthlib==1.3.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# File Processing
Pillow==10.1.0
python-magic==0.4.27
openpyxl==3.1.2
reportlab==4.0.7
python-docx==1.1.0

# API Tools
httpx==0.25.2
aiohttp==3.9.1

# WebSocket
websockets==12.0

# Logging
loguru==0.7.2
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Email
aiosmtplib==3.0.1
jinja2==3.1.2
email-validator==2.1.0

# SMS (optional)
# aliyun-python-sdk-core==2.13.35
# aliyun-dysmsapi==2.2.5

# Cloud Storage
boto3==1.29.7
oss2==2.18.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Development
ipython==8.18.1
ipdb==0.13.13
